Create Python script to start building a public folder from source
This first version parses the contents of the root folder and looks for
`templates/`, `sketches/`.

Every time the script is run, it deletes (if exists) `public/` and
creates a fresh directory where all the built new web pages will be
hosted.

For now, it just creates `public/index.html` to contain links for each
sketch in `sketches/`, and multiple folders `public/{sketch}` with a
custom `index.html` and `sketch.js`. `index.html` is generated using the
filename of the corresponding sketch, and `sketch.js` is just a copy of
the original file.