Major update to logic structure and handling order
Previous version generated a name and id just before requiring them in
one place, and caused no problem or efficiency troubles since were used
only **once**.

However, now we also have `retrieve.js` which generates the value
getters for the project variables, and the `id` of each field is now
required.

Because of this, each field `f` that was an element inside the parsed
`/* ##ginpar */` array, gets a `f['name']` and `f['id']` right after
being converted into a dictionary, as to avoid different values in
different parts of the code, and avoids having to recalculate them.

`retrieve.js` doesn't work as expected yet. It's just a mock for a
future version.